major changes for cleanup
