+++ inc/input.h
@@ -2,10 +2,10 @@
 # define INPUT_H
 
 #include "game.h"
+#include "raycast.h"
 
 # define ROT_ANGLE 3.0
 
-void	close_cub(t_cub *cub);
 void	ft_hook(void *param);
 
 #endif
