+++ src/render.c
@@ -10,74 +10,26 @@
 
 # define BPP sizeof(int32_t)
 
-// Exit the program as failure.
-static void ft_error(void)
-{
-	fprintf(stderr, "%s", mlx_strerror(mlx_errno));
-	exit(EXIT_FAILURE);
-}
-
 // 'Encodes' four individual bytes into an int. Aka sets color to one int
 int get_rgba(int r, int g, int b, int a)
 {
     return (r << 24 | g << 16 | b << 8 | a);
 }
 
-void	render_map_tile(int **map, mlx_image_t *img, int x, int y)
+void	close_cub(t_cub *cub, int err)
 {
-	int	tile_x;
-	int	tile_y;
-	int	tile_color;
-
-	tile_y = 0;
-	if (map[y][x] == WALL)
-				tile_color = get_rgba(0, 0, 0, 255);
-	else if (map[y][x] == FLOOR)
-		tile_color = get_rgba(255, 255, 255, 255);
-	else
-		tile_color = get_rgba(250, 0, 0, 255);
-	while (tile_y < MINI_SCALE)
-	{
-		tile_x = 0;
-		while (tile_x < MINI_SCALE)
-		{
-			mlx_put_pixel(img, (x * MINI_SCALE) + tile_x, (y * MINI_SCALE) + tile_y, tile_color);
-			tile_x++;
-		}
-		tile_y++;
-	}
-	return ;
+	mlx_terminate(cub->mlx);
+	printf("bye\n");
+	free(cub->player);
+	for (int i = 0; i < 5; i++)
+		free(cub->map->m[i]);
+	free(cub->map->m);
+	free(cub->map);
+	free(cub);
+	if (!err)
+		exit(EXIT_SUCCESS);
+	exit(EXIT_FAILURE);
 }
-
-void	render_minimap(t_map *map, mlx_image_t *img)
-{
-	int	y;
-	int	x;
-	
-	y = 0;
-	while (y < map->height)
-	{
-		x = 0;
-		while (x < map->width)
-		{
-			if (map->m[y][x] != ' ')
-				render_map_tile(map->m, img, x, y);
-			x++;
-		}
-		y++;
-	}
-	return ;
-}
-
-//render loop;
-void	render(t_cub *cub)
-{ 
-	raycast(cub);
-	render_minimap(cub->map, cub->img);
-	return ;
- }
-
-
 
 // finds player starting position x and y
 int find_start(t_map_info info, int *x)
@@ -103,7 +55,6 @@ int find_start(t_map_info info, int *x)
 	return (-1);
 }
 
-// image loop
 void	cub3d(t_map_info info)
 {
 	t_cub	*cub;
@@ -116,11 +67,11 @@ void	cub3d(t_map_info info)
 	printf("start[0] is %d start[1] is %d\n", start[0], start[1]);
 	init_cub(cub, info, start);
 	printf("initialized cub\n");	
-	render(cub);
+	raycast(cub);
 	if (!cub->img || (mlx_image_to_window(cub->mlx, cub->img, 0, 0) < 0))
-		ft_error(); //modify later
+		close_cub(cub, 1);
 	mlx_loop_hook(cub->mlx, ft_hook, cub);
 	mlx_loop(cub->mlx);
-	close_cub(cub);
+	close_cub(cub, 0);
 	return ;
 }
